decompose expr.hpp
