+++ include/ast/expressions/expr_entity.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "./expr.hpp"
+
+class expr_entity : public expr {
+public:
+	expr_entity(int begin, int end) : expr(begin, end) {};
+};
