+++ include/ast/expressions/number.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "./expr_entity.hpp"
+
+class expr_number : public expr_entity {
+public:
+	expr_number(token_number* val, int begin, int end) : expr_entity(begin, end), VALUE(val) {};
+
+	token_number* VALUE;
+
+	std::wstring view() {
+		return this->VALUE->view();
+	}
+
+	semantic_node* eval(environment env, const runtime& Runtime);
+};
